requirements.txt
allure-behave==2.9.43
allure-python-commons==2.9.43
Appium-Python-Client==1.2.0
attrs==21.2.0
behave==1.2.6
certifi==2021.5.30
charset-normalizer==2.0.2
docker==5.0.0
idna==3.2
junit-xml==1.9
parse==1.19.0
parse-type==0.5.2
pluggy==0.13.1
pywin32==227
requests==2.26.0
selenium==3.141.0
six==1.16.0
urllib3==1.26.6
websocket-client==1.1.0
